[BugFix] Check the correctness of `UDF_RUNTIME_DIR` before removing it, avoid removing the system files in root directory. (#11535)
(cherry picked from commit 9cada5e5d1f6d1fca40aa905)

+++ bin/start_be.sh
@@ -106,7 +106,9 @@ if [ ! -d $UDF_RUNTIME_DIR ]; then
     mkdir -p ${UDF_RUNTIME_DIR}
 fi
 
-rm -f ${UDF_RUNTIME_DIR}/*
+if [ ! -z ${UDF_RUNTIME_DIR} ]; then
+    rm -f ${UDF_RUNTIME_DIR}/*
+fi
 
 pidfile=$PID_DIR/be.pid
